@@ -28,66 +29,73 @@
   "author": "CosmoX",
   "license": "MIT",
   "devDependencies": {
-    "babel-core": "^6.21.0",
+    "babel-core": "^6.26.3",
     "babel-loader": "^7.0.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
-    "css-loader": "^0.28.4",
-    "electron": "^1.7.6",
-    "electron-builder": "^19.2.7",
-    "eslint": "^4.7.0",
-    "eslint-config-standard": "^10.2.1",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-node": "^5.1.1",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-react": "^7.2.1",
-    "eslint-plugin-standard": "^3.0.1",
-    "file-loader": "^0.11.1",
-    "html-loader": "^0.5.0",
+    "css-loader": "^2.0.0",
+    "electron": "^3.0.0",
+    "electron-builder": "^20.38.2",
+    "eslint": "^5.0.0",
+    "eslint-config-standard": "^12.0.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-node": "^8.0.0",
+    "eslint-plugin-promise": "^4.0.0",
+    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "file-loader": "^2.0.0",
+    "html-loader": "^0.5.5",
     "json-loader": "^0.5.7",
-    "license-checker": "^14.0.0",
-    "node-sass": "^4.5.3",
+    "license-checker": "^24.0.0",
+    "node-sass": "^4.10.0",
     "raw-loader": "^0.5.1",
-    "redux-devtools": "^3.3.1",
-    "sass-loader": "^6.0.5",
-    "style-loader": "^0.18.1",
+    "sass-loader": "^7.0.0",
+    "style-loader": "^0.23.1",
     "text-loader": "0.0.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^3.5.2",
-    "webpack-node-externals": "^1.6.0"
+    "url-loader": "^1.1.2",
+    "webpack": "^4.27.1",
+    "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
-    "autolinker": "^1.4.0",
-    "bluebird": "^3.4.7",
-    "codemirror": "^5.28.0",
-    "electron-context-menu": "^0.9.0",
-    "electron-is-dev": "^0.3.0",
-    "electron-json-storage-sync": "^1.1.0",
-    "electron-localshortcut": "^3.0.0",
-    "electron-updater": "^2.9.3",
-    "electron-window-state": "^4.1.1",
-    "fuse.js": "^3.0.0",
-    "highlight.js": "^9.9.0",
+    "autolinker": "^1.8.1",
+    "bluebird": "^3.5.3",
+    "chart.js": "^1.1.1",
+    "codemirror": "^5.42.0",
+    "electron-context-menu": "^0.10.1",
+    "electron-is-dev": "^1.0.0",
+    "electron-json-storage-sync": "^1.1.1",
+    "electron-localshortcut": "^3.1.0",
+    "electron-updater": "^4.0.0",
+    "electron-window-state": "^5.0.0",
+    "fuse.js": "^3.3.0",
+    "highlight.js": "^9.13.1",
+    "highlightjs-solidity": "^1.0.6",
     "human-readable-time": "^0.3.0",
-    "image-downloader": "^3.2.1",
-    "markdown-it": "^8.3.2",
-    "markdown-it-task-lists": "^2.0.1",
-    "moment": "^2.17.1",
-    "nconf": "^0.8.4",
-    "proxy-agent": "^2.0.0",
-    "react": "^15.5.1",
-    "react-bootstrap": "^0.31.2",
+    "image-downloader": "^3.4.2",
+    "markdown-it": "^8.4.2",
+    "markdown-it-katex": "^2.0.3",
+    "markdown-it-task-lists": "^2.1.1",
+    "moment": "^2.22.2",
+    "nconf": "^0.10.0",
+    "notebookjs": "^0.3.2",
+    "prismjs": "^1.15.0",
+    "proxy-agent": "^3.0.0",
+    "react": "^16.6.3",
+    "react-bootstrap": "^0.32.4",
+    "react-chartjs": "^1.2.0",
     "react-codemirror": "^1.0.0",
-    "react-dom": "^15.5.1",
-    "react-redux": "^5.0.6",
-    "react-split-pane": "^0.1.66",
-    "redux": "^3.7.2",
-    "redux-form": "^7.0.0",
-    "redux-thunk": "^2.1.0",
-    "request": "^2.79.0",
-    "request-promise": "^4.1.1",
-    "twitter-text": "^1.14.3",
-    "winston": "^2.3.0"
+    "react-dom": "^16.6.3",
+    "react-redux": "^5.1.1",
+    "react-split-pane": "^0.1.84",
+    "redux": "^4.0.0",
+    "redux-form": "^7.4.2",
+    "redux-thunk": "^2.3.0",
+    "request": "^2.88.0",
+    "request-promise": "^4.2.2",
+    "twitter-text": "^3.0.0",
+    "valid-filename": "^2.0.1",
+    "webpack-cli": "^3.1.1",
+    "winston": "^2.4.4"
   },
   "build": {
     "appId": "com.cosmox.lepton",
