@@ -30,7 +30,15 @@ export const UPDATE_GIST_NEW_MODAL = 'UPDATE_GIST_NEW_MODAL'
 export const UPDATE_GIST_DELETE_MODAL_STATUS = 'UPDATE_GIST_DELETE_MODAL_STATUS'
 export const UPDATE_PINNED_TAGS_MODAL_STATUS = 'UPDATE_PINNED_TAGS_MODAL_STATUS'
 export const UPDATE_FILE_EXPAND_STATUS = 'UPDATE_FILE_EXPAND_STATUS'
+export const UPDATE_DASHBOARD_MODAL_STATUS = 'UPDATE_DASHBOARD_MODAL_STATUS'
 export const UPDATE_ABOUT_MODAL_STATUS = 'UPDATE_ABOUT_MODAL_STATUS'
+
+export function updateDashboardModalStatus (status) {
+  return {
+    type: UPDATE_DASHBOARD_MODAL_STATUS,
+    payload: status
+  }
+}
 
 export function updateAboutModalStatus (status) {
   return {
@@ -226,7 +234,7 @@ export function fetchSingleGist (oldGist, id) {
       })
       .catch((err) => {
         logger.error('The request has failed: ' + err)
-        Notifier('Sync failed', 'Please check your network condition.')
+        Notifier('Sync failed', 'Please check your network condition. 01')
       })
   }
 }
