@@ -3,13 +3,15 @@
 import HighlightJS from 'highlight.js'
 import MarkdownIt from 'markdown-it'
 import MdTaskList from 'markdown-it-task-lists'
+import MdKatex from 'markdown-it-katex'
 
 // Configure markdown-it
 const Md = MarkdownIt({
-  highlight: function (str, lang) {
+  highlight: (str, lang) => {
     return HighlightJS.highlightAuto(str).value
   }
 })
-Md.use(MdTaskList)
+  .use(MdTaskList)
+  .use(MdKatex, {'throwOnError': false, 'errorColor': ' #cc0000'})
 
 export default Md
