@@ -39,13 +39,13 @@ const template = [
   {
     label: 'View',
     submenu: [
-      {
-        label: 'Reload',
-        accelerator: 'CmdOrCtrl+R',
-        click (item, focusedWindow) {
-          if (focusedWindow) focusedWindow.reload()
-        }
-      },
+      // {
+      //   label: 'Reload',
+      //   accelerator: 'CmdOrCtrl+R',
+      //   click (item, focusedWindow) {
+      //     if (focusedWindow) focusedWindow.reload()
+      //   }
+      // },
       {
         label: 'Toggle Developer Tools',
         accelerator: process.platform === 'darwin' ? 'Alt+Command+I' : 'Ctrl+Shift+I',
