@@ -1,13 +1,27 @@
 module.exports = {
-    'logger': {
-        'level': 'info'
+    "userPanel": {
+        "hideProfilePhoto": false
     },
-    'proxy': {
-        'enable': false,
-        'address': 'socks://localhost:1080'
+    "logger": {
+        "level": "debug"
     },
-    'snippet': {
-        'expanded': true,
-        'newSnippetPrivate': false
+    "proxy": {
+        "enable": false,
+        "address": "socks://localhost:1080"
+    },
+    "snippet": {
+        "sorting": "updated_at",
+        "sortingReverse": true,
+        "expanded": true,
+        "newSnippetPrivate": false
+    },
+    "editor" : {
+        "tabSize": 4
+    },
+    "enterprise": {
+        "enable": false,
+        "host": "",
+        "token": "",
+        "avatarUrl": ""
     }
 }
